+++ LuciusTower_Android/app/src/main/res/drawable/ripple_ffffff.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ripple xmlns:android="http://schemas.android.com/apk/res/android"
+    android:color="#FFFFFF" >
+</ripple>
