+++ LuciusTower_Android/app/src/main/res/layout/item_home_page_tab.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:background="@drawable/ripple_ffffff"
+    android:gravity="center"
+    android:orientation="vertical">
+
+    <androidx.appcompat.widget.AppCompatImageView
+        android:id="@+id/aiv_icon"
+        tools:src="@drawable/ic_male_symbol"
+        android:scaleType="fitCenter"
+        android:layout_width="24dp"
+        android:layout_height="24dp"
+        android:padding="2dp"
+        android:layout_marginTop="3dp" />
+
+    <TextView
+        android:id="@+id/main_bottom_tab_text"
+        tools:ignore="SmallSp"
+        tools:text="test"
+        android:layout_width="wrap_content"
+        android:layout_height="16dp"
+        android:layout_marginTop="2dp"
+        android:layout_marginBottom="3dp"
+        android:gravity="center"
+        android:textColor="#9c9c9c"
+        android:textSize="11sp" />
+
+</androidx.appcompat.widget.LinearLayoutCompat>
